requirements.txt
django
kafka-python
psycopg2
requests